Move: 도메인 객체 위치변경
