+++ src/main/java/nextstep/ladder/README.md
@@ -10,7 +10,6 @@
 - [x] 라인을 생성 할 수 있어야한다. #Line#of
   - [x] 랜덤으로 라인을 생성 할 수 있어야한다.
     - [x] |-----|-----| 모양과 같이 가로 라인이 겹치면 안된다. #Line#validation
-크
 - [ ] 사람의 이름을 받을 수 있다. #LadderInputView#names
 - [ ] 사다리를 출력할 때 사람 이름도 같이 출력한다 #LadderResultView#show
   - [ ] 라인이 있으면 5글자로 출력어야한다.
